backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Flask-Cors==4.0.0
Faker==19.13.0
pymysql==1.1.0
cryptography==41.0.7
redis==5.0.1
